split out Names and State updates
